+++ lang/en/filter_filtercodes.php
@@ -43,6 +43,10 @@ $string['escapebraces'] = 'Escape tags';
 $string['escapebraces_desc'] = 'When this option is checked, you will be able to display FilterCode tags without them being interpreted by this filter by wrapping your tag in [ brackets ]. This can be very useful when creating FilterCodes documentation for the teachers and course creators on your Moodle site.<br><br>Example: [{fullname}] will not display the user\'s full name but display the {fullname} tag instead without the brackets.';
 $string['hidecompletedcourses'] = 'Hide completed courses';
 $string['hidecompletedcourses_desc'] = 'Enable to filter out completed courses in {mycoursesmenu} tag listings.';
+$string['courseteachershowpic'] = 'Show teacher picture';
+$string['courseteachershowpic_desc'] = 'If enabled, will display the teacher\'s profile picture in {courseteachers} tags.';
+$string['courseteacherlinktype'] = 'Teacher link type';
+$string['courseteacherlinktype_desc'] = 'Choose the type of link for the teacher\s link in the {courseteachers} tags.';
 
 $string['sizeb'] = 'B';
 $string['sizekb'] = 'KB';